Create fx_get_relatedness.R
Function to calculate pairwise relatedness. It is based on Hidden Markov Model for IBD estimation. The function use c++ integration in R to speed up the computing time, but it is still slow (about 3 hours for 600 samples). It requires to implement parallelization. Moreover, it only includes a constant error rate model which, and the implementation of a variable error rate model is still in process. In the case of Polyclonal infections, it calculates the least pairwise relatedness between the haplotypes observed in the two samples.