Find drug resistant alleles or haplotypes
Scripts to find drug resistant alleles or haplotypes in a set of samples. It is required to convert it into a function. The script is a little bit slow (~ 1-2 min for 600 samples and 5 drug genes), If possible try to implement c++ integration to apply it at genome scale and for thousand of samples