Random assignment of colors, re-definition of polyclonal samples, and IBD correlation matrix
Random assignment of colors: a palette of 71 colors has been generated, and the assignment of the color palette is consistent across categories on different plots and reports.

re-definition of polyclonal samples: Polyclonal samples can be now defined based on the combination of 4 different criteria: NHetLoci (Number of Heterozygous loci in the sample), Frac_HetLoci (Fraction of heterozygous loci in the samples respect to the number of amplified loci), max_nAlleles (maximum number of alleles found in any of the loci in the sample), and Fws (Within host divergence index).

IBD correlation matrix: The matrix now is build based on the unique combinations of categories between Variable1 and Variable2